requirements.txt
streamlit
openai
azure-storage-blob
azure-ai-formrecognizer
python-dotenv
numpy
langchain-openai
